+++ CHANGELOG.md
@@ -5,6 +5,8 @@ ChangeLog
 0.0.15 (2026-05-01)
 -------------------
 
+- FEATURE: add first cut for a `ase mcp` command which uses the `ase service` under the hood
+
 0.0.14 (2026-05-01)
 -------------------